rosic.cpp: added infrastructure/ includes

+++ Libraries/JUCE/modules/rosic/_third_party/soundtouch/WavFile.cpp
@@ -248,7 +248,7 @@ int WavInFile::read(char *buffer, int maxElems)
         assert(numBytes >= 0);
     }
 
-    numBytes = fread(buffer, 1, numBytes, fptr);
+    numBytes = (int) fread(buffer, 1, numBytes, fptr);
     dataRead += numBytes;
 
     return numBytes;
@@ -290,7 +290,7 @@ int WavInFile::read(short *buffer, int maxElems)
             assert(numBytes >= 0);
         }
 
-        numBytes = fread(buffer, 1, numBytes, fptr);
+        numBytes = (int) fread(buffer, 1, numBytes, fptr);
         dataRead += numBytes;
         numElems = numBytes / 2;
 
@@ -646,7 +646,7 @@ void WavOutFile::writeHeader()
 
     // write the supplemented header in the beginning of the file
     fseek(fptr, 0, SEEK_SET);
-    res = fwrite(&hdrTemp, sizeof(hdrTemp), 1, fptr);
+    res = (int) fwrite(&hdrTemp, sizeof(hdrTemp), 1, fptr);
     if (res != 1)
     {
         throw runtime_error("Error while writing to a wav file.");
@@ -676,7 +676,7 @@ void WavOutFile::write(const char *buffer, int numElems)
     }
     assert(sizeof(char) == 1);
 
-    res = fwrite(buffer, 1, numElems, fptr);
+    res = (int) fwrite(buffer, 1, numElems, fptr);
     if (res != numElems) 
     {
         throw runtime_error("Error while writing to a wav file.");
@@ -717,7 +717,7 @@ void WavOutFile::write(const short *buffer, int numElems)
         memcpy(pTemp, buffer, numElems * 2);
         _swap16Buffer(pTemp, numElems);
 
-        res = fwrite(pTemp, 2, numElems, fptr);
+        res = (int) fwrite(pTemp, 2, numElems, fptr);
 
         delete[] pTemp;
